requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
attrs==24.2.0
braceexpand==0.1.7
cffi==1.17.0
click==8.1.7
discord.py==2.4.0
frozenlist==1.4.1
idna==3.8
import_expression==2.1.0.post1
jishaku @ git+https://github.com/Gorialis/jishaku@0c16ddca174726966704a8f12f6e8f0e26528202
libnacl==2.1.0
multidict==6.0.5
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.0.1
setuptools==72.1.0
tabulate==0.9.0
typing_extensions==4.12.2
wheel==0.43.0
yarl==1.9.6
